+++ quinzena8/astromatch/src/components/Footer/styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+export const FooterContainer = styled.footer`
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+
+    img{
+        height: 9vh;
+    }
+
+    img:hover{
+        height: 10vh;
+    }
+`
