+++ quinzena8/astromatch/src/pages/TelaMatches/TelaMatches.js
@@ -1,40 +1,14 @@
 import React, { useEffect, useState } from 'react'
 import axios from 'axios'
-import styled from 'styled-components'
+import { TelaExibeMatch, CardMatch, ImagemMatch } from './styled.js'
+import { data_Url } from '../../constants/data-url.js'
 
-const TelaExibeMatch = styled.div`
-overflow-y: scroll;
-height: 100%;
-`
-
-const CardMatch = styled.section`
-display: flex;
-padding: 8px;
-align-items: center;
-
-:hover{
-  background-color: lightgray;
-}
-
-:active{
-  background-color: darkgray;
-}
-`
-const ImagemMatch = styled.img`
-  height: 64px;
-  width: 64px;
-  border-radius: 50%;
-  margin-right: 16px;
-
-`
 export default function TelaMatches() {
   const [matches, setMatches] = useState([])
   useEffect(() => {
-    axios.get('https://us-central1-missao-newton.cloudfunctions.net/astroMatch/clara/matches')
+    axios.get(`${data_Url}/matches`)
       .then((resposta) => {
-        console.log(resposta.data.matches)
         setMatches(resposta.data.matches)
-
       }).catch(() => {
         alert("ocorreu um erro")
       })
@@ -42,7 +16,7 @@ export default function TelaMatches() {
 
   const exibeMatches = matches.map((match) => {
     return (
-      <CardMatch>
+      <CardMatch key={match.id}>
         <ImagemMatch src={match.photo} />
         <p>{match.name}</p>
       </CardMatch>
